requirements.txt
# 数据库相关
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23
pymysql==1.1.0

# 向量数据库和机器学习
faiss-cpu==1.12.0
scikit-learn==1.7.1
numpy>=1.25.0

# 可选：高级语义搜索（需要解决依赖冲突）
# chromadb==0.4.15
# sentence-transformers==2.2.2

# 数据处理
pandas==2.0.3
openpyxl==3.1.2

# 文本处理
jieba==0.42.1
markdown==3.5.1

# 配置和工具
python-dotenv==1.0.0
loguru==0.7.2
tqdm==4.66.1
